docs: add benches and document remove runtime
Adds benchamrks covering the remove_resource operation and documents the
O(n) runtime complexity.

This can be optimised to an O(1) operation, and indeed Algorithm 2 from
paper uses a bijective mapping between keys and resources to enable
this. In practice however, this requires an additional Hash bounds on
the resources type, restricting the usability of this implementation
which is optimised for lookups. The current linear search through
resources appears 'fast enough' for most purposes.